requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.3
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
django-environ==0.11.2
python-dateutil==2.9.0
python-json-logger==4.0.0
